+++ package.json
@@ -6,7 +6,7 @@
   "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node src/general/index.js"
+    "start": "node src/index.js"
   },
   "repository": {
     "type": "git",
